dev_requirements.txt
model-mommy==1.2.6
mock==2.0.0
django-test-without-migrations==0.4
